Take the scheduler out of the dispatcher Fixes #6
